jobs/rss-collector/requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
openai>=1.58.1
python-dateutil==2.8.2
